J2-37 Portion calculation & some Recipe and Ingredient Datatypes changed

+++ src/main/java/de/hsba/bi/grp3/recipe/Recipe.java
@@ -30,11 +30,11 @@ public class Recipe {
     @Lob
     private String instruction;
 
-    private BigDecimal prepTime;
+    private Integer prepTime;
 
-    private BigDecimal cookTime;
+    private Integer cookTime;
 
-    private BigDecimal servings;
+    private Integer servings;
 
     @Enumerated
     private Difficulty difficulty;
